requirements.txt
fastmcp
langchain-mcp-adapters
langchain-groq
langchain-core
langgraph
flask
python-dotenv
duckduckgo-search
requests
pillow
asyncio